+++ Samples~/Scripts/HideLabelSample.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+using EditorAttributes;
+
+namespace EditorAttributeSamples
+{
+	[HelpURL("https://editorattributesdocs.readthedocs.io/en/latest/Attributes/hidelabel.html")]
+	public class HideLabelSample : MonoBehaviour
+	{
+		[Header("HideLabel Attribute:")]
+		[SerializeField, HideLabel] private string stringField;
+	}
+}