+++ Scripts/Attributes/Editor/Drawers/DisableFieldDrawer.cs
@@ -9,9 +9,9 @@ namespace EditorAttributes.Editor
 		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
 		{
 			var disableAttribute = attribute as DisableFieldAttribute;
-			var conditionalProperty = GetValidMemberInfo(disableAttribute.ConditionName, property);
+			var conditionalProperty = ReflectionUtility.GetValidMemberInfo(disableAttribute.ConditionName, property);
 
-			GUI.enabled = !GetConditionValue<DisableFieldAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject);
+			GUI.enabled = !GetConditionValue(conditionalProperty, disableAttribute, property);
 
 			DrawProperty(position, property, label);