+++ Scripts/Attributes/Editor/Drawers/EnableFieldDrawer.cs
@@ -9,9 +9,9 @@ namespace EditorAttributes.Editor
 		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
 		{
 			var enableAttribute = attribute as EnableFieldAttribute;
-			var conditionalProperty = GetValidMemberInfo(enableAttribute.ConditionName, property);
+			var conditionalProperty = ReflectionUtility.GetValidMemberInfo(enableAttribute.ConditionName, property);
 
-			GUI.enabled = GetConditionValue<EnableFieldAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject);
+			GUI.enabled = GetConditionValue(conditionalProperty, enableAttribute, property);
 			
 			DrawProperty(position, property, label);