+++ Scripts/Attributes/Editor/Drawers/HideLabelDrawer.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+using UnityEditor;
+
+namespace EditorAttributes.Editor
+{
+    [CustomPropertyDrawer(typeof(HideLabelAttribute))]
+    public class HideLabelDrawer : PropertyDrawerBase
+    {
+    	public override void OnGUI(Rect position, SerializedProperty property, GUIContent label) => DrawProperty(position, property, new GUIContent());
+    }
+}